pyproject.toml
[project]
name = "gerador-relatorio-kiro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "cryptography>=46.0.6",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "kaleido>=1.2.0",
    "langchain>=1.2.15",
    "langchain-core>=1.2.26",
    "langchain-openai>=1.1.12",
    "markdown>=3.10.2",
    "matplotlib>=3.10.8",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "passlib>=1.7.4",
    "pdfplumber>=0.11.9",
    "plotly>=6.6.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.22",
    "python-pptx>=1.0.2",
    "seaborn>=0.13.2",
    "structlog>=25.5.0",
    "uvicorn>=0.43.0",
    "weasyprint>=68.1",
    "xlrd>=2.0.2",
]
